Renames files to index and wrapped location text
